Add Stage 1 Boss Sol Dae & Change Collision Logic

+++ MetalSlug/Line.cpp
@@ -24,8 +24,8 @@ CLine::~CLine()
 
 void CLine::Render(HDC _DC)
 {
-	int scrollX = CScrollMgr::Get_Instance()->Get_ScrollX();
-	int scrollY = CScrollMgr::Get_Instance()->Get_ScrollY();
+	int scrollX = (int)CScrollMgr::Get_Instance()->Get_ScrollX();
+	int scrollY = (int)CScrollMgr::Get_Instance()->Get_ScrollY();
 
 	MoveToEx(_DC, (int)m_tInfo.tLeftPos.fX + scrollX, (int)m_tInfo.tLeftPos.fY + scrollY, nullptr);
 	LineTo(_DC, (int)m_tInfo.tRightPos.fX + scrollX, (int)m_tInfo.tRightPos.fY + scrollY);
